SimonSays ViewController is added
The ViewContoller class for the Simon Says game has been added. The
actual view controller on the storyboard has not yet been added. Also a
“air plane ding” .wav file has been added. The sound is for button
presses. This was found online to test out the class’ audio
functionality - it is to be replaced and at the least added to later
(more .wav files).