Simon Says Theme now matches, added another .wav file, edited storyboard!
CAUTION: Made changes to storyboard. Should only be under the Simon
Says View Controller. But, something to pay attention to before
accepting pull request.

I made the theme match the rest of the app, also added a segue back to
the score board.

In “qlobe.xcodeproj/project.pbxproj” this I believe is just me moving a
file’s location in the project folder. I’m committing it in case the
file ID is referenced in another file I’m committing.

elevatording.wav is another sound I found, not referenced. It’s
probably going to be a temp sound affect until I find other ones or
come up with my own.

Much of the code in SimonSaysViewController is highlighted because
other than making several changes, and adding a “Score Pool”, I
rearranged much of the code for organization to make things easier to
read, and hopefully more understandable?