Bugs Fixed : mostly settings consistency
1. Adjusted cell background color ( to transparent ) when a cell in the
Settings TVC is pressed.
2. Adjusted mute logo to represent it’s current state. Adjusted the
mute logo and audio to also consistently represent the current state
after coming back from the settings. (After coming back from settings
and adjust volume to 0 the logo would always represent that there is
audio playing).
	- noticed that immediately after the loading page the loading audio
continues to play.. this is okay, but if mute is pressed while the
loading audio is playing the loading audio is not affected by mute.
3. Adjusted Simon Says buttons Colors for when they are pressed by the
user.
4. Adjusted volume slider, and rounds slider and text to represent that
of the settings.
5. added “ROUND = 1” to the func LetsBegin in the ViewController, to
reset the round to 1 overtime the game is initiated. - Before after all
rounds were complete, the rounds were never reset, so the scoreboard
was entered every time.