added location feature
currently just testing it out, but added a location option on the
settings page, and corresponding func in the settingsDS class. Also,
tweaked a couple smaller things. Added, a duplicate trivia questions
array. Haven’t done anything with it yet, but it will be used when the
other array has reached it’s end to recover the old array and start the
trivia questions over (refresh).

Also added <KEY><STRING> value pair to the info.plist this allows the
app to ask user for permission to use location when the location switch
in the settings view is turned on.

This is just the start - more to grow from these changes.