Made several changes regarding Location and it's settings
Added Location class for multiple location objects to compare user’s
location to.

Also on the LoadMainViewController page, I’ve added another
communication to parse where location’s can originally be stored. on
app load, the locations are created using the location class and stored
in a global locationOptions class.

I’d like to add a drop down menu in the settings page. If location is
turned on a list of locations the user is enclosed in can be chosen
from. Which then generates a list of trivia questions based off the
location.

Also, I’ve made it so that the location trivia questions are asked, but
once they are finished (for example my “CS BUILDING” location only has
about 5 corresponding questions) so after the 5 questions are asked the
trivia goes on to use other random trivia questions. However at this
point any location trivia questions will be repeated in the original
trivia questions array if the user is to complete the location trivia
array